Fix the netipmid crash issue during BMC booting process
If sending ipmi command during BMC booting process or BMC is busy,
the netipmid service will crash.
The root cause is that the incoming UDP packet process time > 2 seconds,
and triggers the log, but the elapsedSeconds date formatting is incorrect,
it causes the process crash.

Tested:
Execute the below command:
while [ true ] ;do ipmitool -H $BMCIP -U root -P 0penBmc -I lanplus raw 6 1; sleep 2; done
then reboot the BMC, check journal log on BMC:
no netipmid crash log, and find “IPMI command timed out” log

+++ command_table.cpp
@@ -114,7 +114,8 @@ void Table::executeCommand(uint32_t inCommand,
         if (elapsedSeconds > 2s)
         {
             log<level::ERR>("IPMI command timed out",
-                            entry("DELAY=%d", elapsedSeconds.count()));
+                            entry("DELAY=%lld", static_cast<long long int>(
+                                                    elapsedSeconds.count())));
         }
     }
 }
